Define function to perform pingtest on list of IPs
Defined a basic function which pings IP addresses stored in a dictionary
using the bash tool 'ping'. The dictionary also contains a field 'c' to
control the number of times to ping each IP address, and another field
'i' intended to identify between local and external IP addressess. To
be more well-defined later.